Rename table util file to table-migration
