+++ packages/web-components/src/index.ts
@@ -1,2 +1,2 @@
 export { Components, JSX } from './components';
-export { generateTableChildren } from './utils/table';
+export { generateTableChildren } from './utils/table-migration';
